+++ packages/entities/entities-plugins/src/external-links.ts
@@ -1,3 +1,8 @@
 export default {
   jqlang: 'https://jqlang.org/manual/',
+  /**
+   * Plugin conditional execution.
+   * TODO: wait for Doc team provide a new link.
+   */
+  condition: 'https://developer.konghq.com/gateway/routing/expressions/',
 }